split trainer mixins
